Add files via upload
a set of checkboxes on the hobby in forms is done but ERROR if no check mark on all three hobbies.  Might need to add if function to remove the error.  
Also, make it work to go back to the main route("/") when click on go back button from localhost:5000/result.